[FEAT]: Add confirmation-dialog-component. [FEAT]: Implement confirmation-dialog on delete file & directory methods

+++ virtual-files-system-front/src/app/features/list/components/card-file/card-file.component.ts
@@ -7,6 +7,8 @@ import { MatMenuModule } from '@angular/material/menu';
 import { MatButtonModule } from '@angular/material/button';
 import { FilesService } from '../../../../shared/services/files.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmationDialogComponent } from '../../../../shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-card-file',
@@ -18,23 +20,32 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 export class CardFileComponent {
   file = input.required<File>();
 
+  dialog = inject(MatDialog);
   snackBar = inject(MatSnackBar);
   fileService = inject(FilesService);
 
   fileName = computed(() => this.file().fileName);
 
   onClickDelete(fileId: number): void {
-    this.fileService.deleteFile(fileId).subscribe({
-      next: () => {
-        this.snackBar.open("Arquivo deletado!", "Fechar");
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: { name: this.fileName() }
+    });
 
-        setTimeout(() => {
-            location.reload();
-        }, 1000);
-      },
-      error: (error) => {
-        console.error("Erro ao deletar o diretório:", error);
-        this.snackBar.open("Erro ao deletar o diretório!", "Fechar");
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.fileService.deleteFile(fileId).subscribe({
+          next: () => {
+            this.snackBar.open("Arquivo deletado!", "Fechar");
+
+            setTimeout(() => {
+              location.reload();
+            }, 1000);
+          },
+          error: (error) => {
+            console.error("Erro ao deletar o diretório:", error);
+            this.snackBar.open("Erro ao deletar o diretório!", "Fechar");
+          }
+        });
       }
     });
   }
